Factor out registry part of testing
This makes the registry helpers separate to the rest of the suite, and
adds a test for the tags listing helper itself.
